Develop

+++ app/Http/Controllers/DivisionsController.php
@@ -15,8 +15,9 @@ class DivisionsController extends Controller
     public function index()
     {
         //
+        $this->authorize('viewAny', Divisions::class);
         $divisions = divisions::all();
-        return view('divisions.index')->with('divisions', $divisions);
+        return view('divisions.index')->with('divisions', $divisions)->with('scripts',true);
     }
 
     /**
@@ -39,6 +40,7 @@ class DivisionsController extends Controller
     public function store(Request $request)
     {
         //
+        $this->authorize('create', Divisions::class);
         $validateData = $request->validate([
             'nama_divisi' => 'required'
         ]);
@@ -47,9 +49,8 @@ class DivisionsController extends Controller
         $divisions->nama_divisi = $validateData['nama_divisi'];
 
         $divisions->save();
+        $request->session()->flash('success', 'Data Divisi berhasil di simpan');
         return redirect()->route('divisions.index');
-        
-        
     }
 
     /**
@@ -85,12 +86,13 @@ class DivisionsController extends Controller
     public function update(Request $request, Divisions $division)
     {
         //
+        $this->authorize('update', Divisions::class);
         $validateData =$request->validate([
             'nama_divisi' => 'required'
         ]);
 
         Divisions::where('id', $division->id)->update($validateData);
-        $request->session()->flash('info', 'Data Penyelenggara berhasil di ubah');
+        $request->session()->flash('success', 'Data Penyelenggara berhasil di edit');
         return redirect()->route('divisions.index');
     
     }
@@ -104,7 +106,8 @@ class DivisionsController extends Controller
     public function destroy(Divisions $division)
     {
         //
+        $this->authorize('delete', Divisions::class);
         $division->delete();
-        return redirect()->route('divisions.index')->with('info', "Data divisi $division->nama_divisi berhasil dihapus");
+        return redirect()->route('divisions.index')->with('success', "Data divisi $division->nama_divisi berhasil dihapus");
     }
 }
