+++ app/Http/Controllers/PeriodsController.php
@@ -15,8 +15,9 @@ class PeriodsController extends Controller
     public function index()
     {
         //
+        $this->authorize('viewAny', Periods::class);
         $periods = Periods::all();
-        return view('periods.index')->with('periods', $periods);
+        return view('periods.index')->with('periods', $periods)->with('scripts', true);
     }
 
     /**
@@ -39,16 +40,16 @@ class PeriodsController extends Controller
     public function store(Request $request)
     {
         //
+        $this->authorize('create', Periods::class);
         $validateData = $request->validate([
-            'tahun_akademik' => 'required',
-            'semester' => 'required'
+            'tahun_akademik' => 'required'
         ]);
 
         $periods = new Periods();
         $periods->tahun_akademik = $validateData['tahun_akademik'];
-        $periods->semester = $validateData['semester'];
 
         $periods->save();
+        $request->session()->flash('success', 'Data Periode berhasil di simpan');
         return redirect()->route('periods.index');
     }
 
@@ -85,13 +86,13 @@ class PeriodsController extends Controller
     public function update(Request $request, Periods $period)
     {
         //
+        $this->authorize('update', Periods::class);
         $validateData = $request->validate([
-            'tahun_akademik' => 'required',
-            'semester' => 'required'
+            'tahun_akademik' => 'required'
         ]);
 
         Periods::where('id', $period->id)->update($validateData);
-        $request->session()->flash('info', 'Data Periode berhasil di ubah');
+        $request->session()->flash('success', 'Data Periode berhasil di edit');
         return redirect()->route('periods.index');
     }
 
@@ -104,7 +105,8 @@ class PeriodsController extends Controller
     public function destroy(Periods $period)
     {
         //
+        $this->authorize('delete', Periods::class);
         $period->delete();
-        return redirect()->route('periods.index')->with('info', "Data divisi $period->tahun_akademik berhasil dihapus");
+        return redirect()->route('periods.index')->with('success', "Data Periode $period->tahun_akademik berhasil dihapus");
     }
 }
