+++ app/Http/Controllers/PositionsController.php
@@ -15,8 +15,9 @@ class PositionsController extends Controller
     public function index()
     {
         //
+        $this->authorize('viewAny', Positions::class);
         $positions = positions::all();
-        return view('positions.index')->with('positions', $positions);
+        return view('positions.index')->with('positions', $positions)->with('scripts',true);
     }
 
     /**
@@ -39,6 +40,7 @@ class PositionsController extends Controller
     public function store(Request $request)
     {
         //
+        $this->authorize('create', Positions::class);
         $validateData = $request->validate([
             'nama_jabatan' => 'required'
         ]);
@@ -47,6 +49,7 @@ class PositionsController extends Controller
         $positions->nama_jabatan = $validateData['nama_jabatan'];
 
         $positions->save();
+        $request->session()->flash('success', "Data Jabatan $positions->nama_jabatan berhasil di simpan");
         return redirect()->route('positions.index');
     }
 
@@ -83,12 +86,13 @@ class PositionsController extends Controller
     public function update(Request $request, Positions $position)
     {
         //
+        $this->authorize('update', Positions::class);
         $validateData =$request->validate([
             'nama_jabatan' => 'required'
         ]);
 
         Positions::where('id', $position->id)->update($validateData);
-        $request->session()->flash('info', 'Data Posisi berhasil di ubah');
+        $request->session()->flash('success', 'Data Jabatan berhasil di ubah');
         return redirect()->route('positions.index');
     }
 
@@ -101,7 +105,8 @@ class PositionsController extends Controller
     public function destroy(Positions $position)
     {
         //
+        $this->authorize('delete', Positions::class);
         $position->delete();
-        return redirect()->route('positions.index')->with('info', "Data divisi $position->nama_jabatan berhasil dihapus");
+        return redirect()->route('positions.index')->with('success', "Data Jabatan $position->nama_jabatan berhasil dihapus");
     }
 }
