+++ app/Http/Controllers/StudyProgramsController.php
@@ -15,8 +15,9 @@ class StudyProgramsController extends Controller
     public function index()
     {
         //
+        $this->authorize('viewAny', Study_programs::class);
         $study_programs = study_programs::all();
-        return view('study_programs.index')->with('study_programs', $study_programs);
+        return view('study_programs.index')->with('study_programs', $study_programs)->with('scripts',true);
     }
 
     /**
@@ -39,6 +40,7 @@ class StudyProgramsController extends Controller
     public function store(Request $request)
     {
         //
+        $this->authorize('create', Study_programs::class);
         $validateData = $request->validate([
             'nama_prodi' => 'required'
         ]);
@@ -47,6 +49,7 @@ class StudyProgramsController extends Controller
         $study_programs->nama_prodi = $validateData['nama_prodi'];
 
         $study_programs->save();
+        $request->session()->flash('success', "Data Program Studi $study_programs->nama_prodi berhasil di simpan");
         return redirect()->route('study_programs.index');
     }
 
@@ -83,12 +86,13 @@ class StudyProgramsController extends Controller
     public function update(Request $request, Study_programs $study_program)
     {
         //
+        $this->authorize('update', Study_programs::class);
         $validateData =$request->validate([
             'nama_prodi' => 'required'
         ]);
 
         Study_programs::where('id', $study_program->id)->update($validateData);
-        $request->session()->flash('info', 'Data Program Studi berhasil di ubah');
+        $request->session()->flash('success', 'Data Program Studi berhasil di ubah');
         return redirect()->route('study_programs.index');
     }
 
@@ -101,7 +105,8 @@ class StudyProgramsController extends Controller
     public function destroy(Study_programs $study_program)
     {
         //
+        $this->authorize('delete', Study_programs::class);
         $study_program->delete();
-        return redirect()->route('study_programs.index')->with('info', "Data divisi $study_program->nama_prodi berhasil dihapus");
+        return redirect()->route('study_programs.index')->with('success', "Data Program Studi $study_program->nama_prodi berhasil dihapus");
     }
 }
